+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Add winding check in map_mesh function.
+
 ### Changed
 
 * Added meshpatter unify_cycles from COMPAS, LIBIGL has only triangle BFS.